refactor wounds into numbers
