+++ src/constants/woundTable.ts
@@ -1,93 +1,85 @@
 import { diceType } from './dice';
 import { woundIntensityType } from './woundIntensity';
-import {
-  rien,
-  sonne,
-  legere,
-  grave,
-  critique,
-  tueNet,
-  woundResultsType,
-} from './woundResults';
+import { woundResultsType } from './woundResults';
 
 export const woundTable: WoundTableType = {
   1: {
-    [-1]: rien,
-    0: sonne,
-    1: sonne,
-    2: legere,
-    3: legere,
-    4: legere,
-    5: grave,
-    6: grave,
-    7: grave,
-    8: critique,
-    9: critique,
+    [-1]: 0,
+    0: 1,
+    1: 1,
+    2: 2,
+    3: 2,
+    4: 2,
+    5: 3,
+    6: 3,
+    7: 3,
+    8: 4,
+    9: 4,
   },
   2: {
-    [-1]: sonne,
-    0: sonne,
-    1: legere,
-    2: legere,
-    3: grave,
-    4: grave,
-    5: grave,
-    6: critique,
-    7: critique,
-    8: critique,
-    9: tueNet,
+    [-1]: 1,
+    0: 1,
+    1: 2,
+    2: 2,
+    3: 3,
+    4: 3,
+    5: 3,
+    6: 4,
+    7: 4,
+    8: 4,
+    9: 5,
   },
   3: {
-    [-1]: sonne,
-    0: legere,
-    1: legere,
-    2: grave,
-    3: grave,
-    4: critique,
-    5: critique,
-    6: critique,
-    7: tueNet,
-    8: tueNet,
-    9: tueNet,
+    [-1]: 1,
+    0: 2,
+    1: 2,
+    2: 3,
+    3: 3,
+    4: 4,
+    5: 4,
+    6: 4,
+    7: 5,
+    8: 5,
+    9: 5,
   },
   4: {
-    [-1]: legere,
-    0: legere,
-    1: grave,
-    2: grave,
-    3: critique,
-    4: critique,
-    5: tueNet,
-    6: tueNet,
-    7: tueNet,
-    8: tueNet,
-    9: tueNet,
+    [-1]: 2,
+    0: 2,
+    1: 3,
+    2: 3,
+    3: 4,
+    4: 4,
+    5: 5,
+    6: 5,
+    7: 5,
+    8: 5,
+    9: 5,
   },
   5: {
-    [-1]: legere,
-    0: grave,
-    1: critique,
-    2: critique,
-    3: tueNet,
-    4: tueNet,
-    5: tueNet,
-    6: tueNet,
-    7: tueNet,
-    8: tueNet,
-    9: tueNet,
+    [-1]: 2,
+    0: 3,
+    1: 4,
+    2: 4,
+    3: 5,
+    4: 5,
+    5: 5,
+    6: 5,
+    7: 5,
+    8: 5,
+    9: 5,
   },
   6: {
-    [-1]: legere,
-    0: grave,
-    1: critique,
-    2: critique,
-    3: tueNet,
-    4: tueNet,
-    5: tueNet,
-    6: tueNet,
-    7: tueNet,
-    8: tueNet,
-    9: tueNet,
+    [-1]: 2,
+    0: 3,
+    1: 4,
+    2: 4,
+    3: 5,
+    4: 5,
+    5: 5,
+    6: 5,
+    7: 5,
+    8: 5,
+    9: 5,
   },
 };
